fix(api): Fix completed message JSON parse — Duration int→float64
ROOT CAUSE FOUND: Python worker sends output.duration as float (7036.69)
but Go struct TranscodeOutputData.Duration was int. JSON unmarshal fails
silently → completed message DROPPED → DB never updated via NATS.

This is why progress 10-95% worked (no output field) but completed always failed.

Fix: Duration int → float64 in TranscodeOutputData + ProgressData

+++ _gofiber_starter/domain/ports/messaging.go
@@ -68,7 +68,7 @@ type ProgressData struct {
 	FileSize int64
 
 	// Transcode output (from new worker completed message)
-	Duration     int
+	Duration     float64
 	DiskUsage    int64
 	QualitySizes map[string]int64
 }
